requirements.txt
flask
tensorflow
keras
numpy
pillow
tqdm
nltk
python-dotenv